# SIH26168-DeadReckoning/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH26168 Dead Reckoning API",
    "description": "Smartphone-only GNSS-blackout dead reckoning: ingest a raw IMU(+GPS) trace, then fetch the naive strapdown/ZUPT baseline trajectory, the ML-residual-corrected trajectory, its duration-bucketed confidence radius, and headline evaluation metrics (ATE/RTE/CEP/drift-rate).\n\n**Baseline, corrected, and uncertainty are currently mock data** -- deterministic per trace, stable across repeated calls, generated by `backend/services/pipeline.py` -- while `physics-baseline/` and `ml-residual/` are still in development. Every response's `source` field says which pipeline produced it, so client code can rely on that field rather than tracking the swap-over out of band.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Liveness check",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/traces": {
      "post": {
        "tags": [
          "traces"
        ],
        "summary": "Ingest a raw IMU(+GPS) trace",
        "description": "Accepts a full trace as one JSON body (not a stream) -- simplest\ncontract for a hackathon demo where traces are recorded on-device and\nuploaded once the walk/drive is done.\n\nSamples are sorted by `timestamp` server-side, so client ordering\ndoesn't matter. `time_unit: \"auto\"` (the default) detects seconds vs.\nmilliseconds vs. nanoseconds the same way `physics-baseline/strapdown.py`\ndoes.\n\nBaseline/corrected/uncertainty are NOT computed here -- they're mock-\ngenerated on demand by the GET endpoints below, deterministically from\nthe trace's duration, so this endpoint stays fast regardless of trace\nlength. Evaluation metrics (ATE/RTE/CEP/drift-rate) ARE computed here\nand stored, since they summarize a completed trace once.",
        "operationId": "ingest_trace_traces_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceIngestRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceSummary"
                }
              }
            }
          },
          "422": {
            "description": "Malformed samples (see detail)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "traces"
        ],
        "summary": "List ingested traces",
        "operationId": "list_traces_traces_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/traces/{trace_id}": {
      "get": {
        "tags": [
          "traces"
        ],
        "summary": "Get trace metadata",
        "operationId": "get_trace_traces__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceSummary"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/traces/{trace_id}/baseline": {
      "get": {
        "tags": [
          "trajectory"
        ],
        "summary": "Naive strapdown/ZUPT trajectory (no ML correction)",
        "description": "The uncorrected dead-reckoned path -- what strapdown mechanization +\nZUPT alone produces, before the ML residual correction. Currently mock\ndata (see `services/pipeline.py`); will be replaced by real\n`physics-baseline/strapdown.py` output once that stage is wired in.",
        "operationId": "get_baseline_traces__trace_id__baseline_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/traces/{trace_id}/corrected": {
      "get": {
        "tags": [
          "trajectory"
        ],
        "summary": "Baseline + ML residual correction",
        "description": "Baseline trajectory with the LightGBM residual correction applied\n(dossier Section 4.2/7). Currently mock data that simulates a partial\n(not perfect) drift correction, so this endpoint is visibly closer to\nground truth than /baseline but not identical to it -- the same\nqualitative relationship the real ml-residual model will have.",
        "operationId": "get_corrected_traces__trace_id__corrected_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/traces/{trace_id}/uncertainty": {
      "get": {
        "tags": [
          "trajectory"
        ],
        "summary": "Duration-bucketed confidence radius",
        "description": "Confidence radius per point, bucketed by elapsed time since the last\nGNSS fix (dossier Section 4.3 / jury Q8) -- widens with blackout\nduration, nothing else. Bucket radii are a mock calibration for now\n(`services/pipeline.UNCERTAINTY_BUCKETS`); real values will come from\n`ml-residual/uncertainty.py`'s cross-validated CEP-90-per-bucket fit.\n`blackout_duration_s` per point is derived from this trace's own\ningested GPS-fix timestamps, so it reflects real trace metadata even\nthough the radius values themselves are mock.",
        "operationId": "get_uncertainty_traces__trace_id__uncertainty_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UncertaintyResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/traces/{trace_id}/evaluation": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "ATE / RTE / CEP-50 / CEP-90 / drift-rate for a trace",
        "operationId": "get_evaluation_traces__trace_id__evaluation_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationMetrics"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "EvaluationMetrics": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "ate_m": {
            "type": "number",
            "title": "Ate M",
            "description": "Absolute Trajectory Error: RMS deviation, corrected vs ground truth, meters."
          },
          "rte_m": {
            "type": "number",
            "title": "Rte M",
            "description": "Relative Trajectory Error: mean drift over rte_window_s sub-windows, meters."
          },
          "rte_window_s": {
            "type": "number",
            "title": "Rte Window S"
          },
          "cep50_m": {
            "type": "number",
            "title": "Cep50 M",
            "description": "Radius within which true position falls 50% of the time."
          },
          "cep90_m": {
            "type": "number",
            "title": "Cep90 M",
            "description": "Radius within which true position falls 90% of the time."
          },
          "drift_rate_pct": {
            "type": "number",
            "title": "Drift Rate Pct",
            "description": "Final position error as a percentage of total distance travelled."
          },
          "distance_m": {
            "type": "number",
            "title": "Distance M",
            "description": "Total distance travelled (mock ground-truth path), meters."
          },
          "computed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Computed At"
          },
          "source": {
            "type": "string",
            "enum": [
              "mock",
              "evaluation"
            ],
            "title": "Source",
            "default": "mock"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "ate_m",
          "rte_m",
          "rte_window_s",
          "cep50_m",
          "cep90_m",
          "drift_rate_pct",
          "distance_m",
          "computed_at"
        ],
        "title": "EvaluationMetrics"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IMUSample": {
        "properties": {
          "timestamp": {
            "type": "number",
            "title": "Timestamp",
            "description": "Raw device timestamp, in the trace's declared time_unit."
          },
          "acc_x": {
            "type": "number",
            "title": "Acc X",
            "description": "Raw accelerometer X, m/s^2 (specific force, includes gravity)."
          },
          "acc_y": {
            "type": "number",
            "title": "Acc Y",
            "description": "Raw accelerometer Y, m/s^2."
          },
          "acc_z": {
            "type": "number",
            "title": "Acc Z",
            "description": "Raw accelerometer Z, m/s^2."
          },
          "gyro_x": {
            "type": "number",
            "title": "Gyro X",
            "description": "Gyroscope X, rad/s."
          },
          "gyro_y": {
            "type": "number",
            "title": "Gyro Y",
            "description": "Gyroscope Y, rad/s."
          },
          "gyro_z": {
            "type": "number",
            "title": "Gyro Z",
            "description": "Gyroscope Z, rad/s."
          },
          "mag_x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mag X",
            "description": "Magnetometer X, microtesla. Omit entirely if the device has no magnetometer."
          },
          "mag_y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mag Y"
          },
          "mag_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mag Z"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat",
            "description": "GPS latitude, degrees. Present only on samples with a GNSS fix."
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon",
            "description": "GPS longitude, degrees."
          },
          "alt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt",
            "description": "GPS altitude, meters."
          },
          "gps_accuracy_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Accuracy M",
            "description": "GPS horizontal accuracy estimate, meters."
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "acc_x",
          "acc_y",
          "acc_z",
          "gyro_x",
          "gyro_y",
          "gyro_z"
        ],
        "title": "IMUSample",
        "description": "One raw sample as captured by the phone's sensor-logging app.\n\n`timestamp` is the device's raw clock reading, in whatever unit the\ntrace declares (see `TraceIngestRequest.time_unit`) -- seconds,\nmilliseconds, or nanoseconds since some device-local epoch. It is\n*not* required to be Unix time.",
        "example": {
          "acc_x": 0.12,
          "acc_y": -0.05,
          "acc_z": 9.79,
          "alt": 216.0,
          "gps_accuracy_m": 4.5,
          "gyro_x": 0.001,
          "gyro_y": -0.002,
          "gyro_z": 0.0005,
          "lat": 28.6139,
          "lon": 77.209,
          "mag_x": 22.1,
          "mag_y": -5.4,
          "mag_z": 41.8,
          "timestamp": 1735689600.123
        }
      },
      "TraceIngestRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id",
            "description": "Free-form identifier for the recording device/phone."
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Human-readable note, e.g. 'metro tunnel walk 1'."
          },
          "time_unit": {
            "type": "string",
            "enum": [
              "auto",
              "s",
              "ms",
              "ns"
            ],
            "title": "Time Unit",
            "description": "Unit of every sample's `timestamp`. 'auto' detects s/ms/ns by magnitude.",
            "default": "auto"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/IMUSample"
            },
            "type": "array",
            "minItems": 2,
            "title": "Samples",
            "description": "Raw IMU(+GPS) samples, any order (server sorts by timestamp)."
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "samples"
        ],
        "title": "TraceIngestRequest"
      },
      "TraceListResponse": {
        "properties": {
          "traces": {
            "items": {
              "$ref": "#/components/schemas/TraceSummary"
            },
            "type": "array",
            "title": "Traces"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "traces",
          "count"
        ],
        "title": "TraceListResponse"
      },
      "TraceSummary": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "time_unit": {
            "type": "string",
            "enum": [
              "s",
              "ms",
              "ns"
            ],
            "title": "Time Unit"
          },
          "start_timestamp": {
            "type": "number",
            "title": "Start Timestamp",
            "description": "First sample's raw timestamp (trace's time_unit)."
          },
          "end_timestamp": {
            "type": "number",
            "title": "End Timestamp",
            "description": "Last sample's raw timestamp (trace's time_unit)."
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          },
          "has_mag": {
            "type": "boolean",
            "title": "Has Mag"
          },
          "has_gps": {
            "type": "boolean",
            "title": "Has Gps"
          },
          "status": {
            "type": "string",
            "const": "ready",
            "title": "Status",
            "description": "Mock pipeline is synchronous: a trace is 'ready' as soon as it's ingested.",
            "default": "ready"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "device_id",
          "label",
          "created_at",
          "time_unit",
          "start_timestamp",
          "end_timestamp",
          "duration_s",
          "sample_count",
          "has_mag",
          "has_gps"
        ],
        "title": "TraceSummary"
      },
      "TrajectoryPoint": {
        "properties": {
          "t": {
            "type": "number",
            "title": "T",
            "description": "Seconds elapsed since trace start."
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp",
            "description": "Raw timestamp, trace's time_unit (same clock as the ingested samples)."
          },
          "pos_x": {
            "type": "number",
            "title": "Pos X",
            "description": "Local ENU East position, meters, origin at trace start."
          },
          "pos_y": {
            "type": "number",
            "title": "Pos Y",
            "description": "Local ENU North position, meters."
          },
          "pos_z": {
            "type": "number",
            "title": "Pos Z",
            "description": "Local ENU Up position, meters."
          },
          "velocity": {
            "type": "number",
            "title": "Velocity",
            "description": "Speed, m/s."
          },
          "orientation": {
            "type": "number",
            "title": "Orientation",
            "description": "Heading, degrees, 0-360, 0=North."
          }
        },
        "type": "object",
        "required": [
          "t",
          "timestamp",
          "pos_x",
          "pos_y",
          "pos_z",
          "velocity",
          "orientation"
        ],
        "title": "TrajectoryPoint"
      },
      "TrajectoryResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "mock",
              "physics-baseline",
              "physics-baseline+ml-residual"
            ],
            "title": "Source"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrajectoryPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "source",
          "generated_at",
          "points"
        ],
        "title": "TrajectoryResponse"
      },
      "UncertaintyBucket": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label",
            "description": "e.g. '0-30s', '30-90s', '90s+'."
          },
          "lo_s": {
            "type": "number",
            "title": "Lo S"
          },
          "hi_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hi S",
            "description": "null means unbounded (the last bucket)."
          },
          "radius_m": {
            "type": "number",
            "title": "Radius M",
            "description": "CEP-`coverage` confidence radius for this blackout-duration bucket."
          }
        },
        "type": "object",
        "required": [
          "label",
          "lo_s",
          "radius_m"
        ],
        "title": "UncertaintyBucket"
      },
      "UncertaintyPoint": {
        "properties": {
          "t": {
            "type": "number",
            "title": "T",
            "description": "Seconds elapsed since trace start."
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "blackout_duration_s": {
            "type": "number",
            "title": "Blackout Duration S",
            "description": "Seconds since the most recent GNSS fix at or before t (or since trace start, if none yet)."
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "radius_m": {
            "type": "number",
            "title": "Radius M"
          }
        },
        "type": "object",
        "required": [
          "t",
          "timestamp",
          "blackout_duration_s",
          "bucket",
          "radius_m"
        ],
        "title": "UncertaintyPoint"
      },
      "UncertaintyResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "mock",
              "ml-residual"
            ],
            "title": "Source"
          },
          "coverage": {
            "type": "number",
            "title": "Coverage",
            "description": "CEP coverage fraction the radii represent, e.g. 0.9 = CEP-90.",
            "default": 0.9
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/UncertaintyBucket"
            },
            "type": "array",
            "title": "Buckets"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/UncertaintyPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "source",
          "generated_at",
          "buckets",
          "points"
        ],
        "title": "UncertaintyResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}